+++ app/Http/Controllers/ItemController.php
@@ -75,6 +75,8 @@ class ItemController extends Controller
 
         $pagination = UtilityPagination::init($request->path(), $total)
             ->setParameters($this->collection_parameters)
+            ->setSortParameters($sort_fields)
+            ->setSearchParameters($search_conditions)
             ->paging();
 
         $items = (new Item())->paginatedCollection(
@@ -420,8 +422,8 @@ class ItemController extends Controller
             (new SubCategory())->paginatedCollection($this->collection_parameters['category'])->map(
                 function ($sub_category)
                 {
-                    $this->get_parameters['subcategory']['allowed_values'][$this->hash->encode('sub_category', $sub_category->id)] = [
-                        'value' => $this->hash->encode('sub_category', $sub_category->id),
+                    $this->get_parameters['subcategory']['allowed_values'][$this->hash->encode('subcategory', $sub_category->id)] = [
+                        'value' => $this->hash->encode('subcategory', $sub_category->id),
                         'name' => $sub_category->name,
                         'description' => trans('item/allowed-values.description-prefix-subcategory') .
                             $sub_category->name . trans('item/allowed-values.description-suffix-subcategory')